Moved individuals_from_obj_values_dummy to test_utils module
