+++ optirustic/src/core/mod.rs
@@ -12,6 +12,7 @@ mod error;
 mod individual;
 mod objective;
 mod problem;
+#[cfg(test)]
 pub(crate) mod test_utils;
 pub mod utils;
 mod variable;
